hllm/requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-llms-openai-like
fastapi[standard]
black